remove pin count check in evict_page(); unpin() moved to evict_all_pages()
The check should not be there, as it will cause trouble in cases which require evict_page() without closing file. In such cases, the header frame has always pin_count > 0,
as the file is not closed and the header frame should be kept. One such scenario is when index file is sync()-ed.
unpin() is moved as the extra pin() is caused by find_list() within evict_all_pages(). The original position will cause trouble when evict_page() is called separately,
for example when the memory is not enough.